facial_recognition_model/requirements.txt
dlib==19.20.0
scikit_image==0.17.2
Flask==1.1.2
Werkzeug==1.0.1
scipy==1.4.1
pyodbc==4.0.30
imutils==0.5.3
gevent==20.6.2
numpy==1.18.5
requests==2.24.0
urllib3==1.25.9
tf_nightly==2.4.0.dev20200730
opencv_python==4.3.0.36
certifi==2020.6.20
Pillow==7.2.0
tensorflow==2.3.0
text_to_image==0.0.5
scikit-image==0.17.2
azure-common==1.1.25
azure-iot-device==2.1.4
azure-storage-blob==2.1.0
azure-storage-common==2.1.0